requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-modeltranslation==0.19.10
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
flower==2.0.1
humanize==4.11.0
inflection==0.5.1
kombu==5.4.2
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
redis==5.0.1
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
django-parler==2.3
django-rosetta==0.10.1